pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "minimax-optimization"
version = "0.1.0"
description = "Christensen-consistent minimax robustness under selective observation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Andrew Watson" }
]
dependencies = []

[project.optional-dependencies]
hf = [
  "torch>=2.2",
  "transformers>=4.40",
]
wilds = [
  "accelerate>=0.28",
  "datasets>=2.18",
  "pandas<3",
  "pyyaml>=6.0",
  "scikit-learn>=1.4",
  "torch>=2.2",
  "transformers>=4.40",
  "wilds>=2.0.0",
]
ag = [
  "ag-survival-sim @ git+https://github.com/adawgwats/ag-survival-sim.git",
]

[project.scripts]
minimax-monte-carlo = "minimax_core.monte_carlo:main"
minimax-gradient-validation = "minimax_core.gradient_validation:main"
minimax-compare-baselines = "minimax_core.comparison:main"
minimax-ag-benchmark = "minimax_core.ag_benchmark:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: requires optional external packages or a real DSSAT installation",
]

[tool.setuptools.packages.find]
include = ["minimax_core*", "minimax_hf*"]
